pyproject.toml
[project]
name = "marc-pandas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lxml>=6.0.2",
    "pandas>=2.3.3",
    "pandera>=0.27.1",
    "python-dotenv>=1.2.1",
    "tabulate>=0.9.0",
]
